backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
geopandas==0.14.4
pandas==2.2.2
shapely==2.0.4
scipy==1.13.0
scikit-learn==1.5.0
pyproj==3.6.1
fiona==1.9.6
requests==2.32.2
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
